pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "dicentra"
version = "2.0.0"
description = "A modern, modular DICOM viewer for clinical imaging workflows"
readme = "README.md"
license = { file = "LICENSE" }
authors = [{ name = "Kareem", email = "" }]
keywords = ["DICOM", "medical imaging", "radiology", "PyQt5", "viewer"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: X11 Applications :: Qt",
    "Intended Audience :: Healthcare Industry",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering :: Medical Science Apps.",
    "Topic :: Scientific/Engineering :: Visualization",
]
requires-python = ">=3.9"
dependencies = [
    "pydicom>=2.4.0",
    "numpy>=1.24.0",
    "scipy>=1.10.0",
    "PyQt5>=5.15.0",
    "matplotlib>=3.7.0",
    "Pillow>=9.0.0",
    "imageio>=2.28.0",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.1.0",
    "black>=24.0.0",
    "pre-commit>=3.5.0",
]

[project.urls]
Homepage = "https://github.com/your-username/dicentra"
Repository = "https://github.com/your-username/dicentra"
"Bug Tracker" = "https://github.com/your-username/dicentra/issues"

[project.scripts]
dicentra = "main:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["app*", "config*"]

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.ruff]
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "W", "I"]
ignore = [
    "E501",   # line too long — black handles length
    "E701",   # multiple statements on one line (colon) — intentional style
    "E702",   # multiple statements on one line (semicolon) — intentional style
    "E741",   # ambiguous variable name — false positives
    "E402",   # module level import not at top — needed for warnings filter in main.py
    "F401",   # imported but unused — re-exports in __init__.py are intentional
    "I001",   # import block unsorted — handled separately
]

[tool.ruff.lint.per-file-ignores]
"app/utils/__init__.py" = ["F401"]
"app/data/__init__.py" = ["F401"]
"app/logic/__init__.py" = ["F401"]
"tests/*" = ["E402", "F401", "I001"]
"examples/*" = ["E402", "I001"]
